feat: Create directories if they don't exist in SkinService constructor
The code changes in the SkinService constructor add logic to create the necessary directories if they don't already exist. This ensures that the "files/skins" directory is available for storing skin uploads. This feature enhancement improves the functionality of the application by automatically creating the required directories, reducing the chance of errors during skin uploads.